Resolve mismatch between database and java model
Changed userId datatype from Long to Integer to match SQL script and configured Hibernate naming strategy so the model does not automatically change column names from camelCase to snake case (e.g., userId won't get changed to user_Id)

+++ Backend/acmeplex/src/main/java/org/example/acmeplex/repository/RegisteredUserRepository.java
@@ -3,7 +3,7 @@ package org.example.acmeplex.repository;
 import org.example.acmeplex.model.RegisteredUser;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RegisteredUserRepository extends JpaRepository<RegisteredUser, Long> {
+public interface RegisteredUserRepository extends JpaRepository<RegisteredUser, Integer> {
     RegisteredUser findByUsername(String username);
     boolean existsByUsername(String username);
     boolean existsByEmail(String email);
